Add IV to the key-material
In order to be able to decrypt the message, it is necessary to have
the iv. The iv is generated during the encryption and will be stored
in the key-part of the decryption-url. The key can be seperated from
the iv by the '&'-sign. The '&'-sign is not part of the base64-charset.